requirements/base.txt
django>=2.2.10,<3.0
djangorestframework>=3.9.0,<4.0
psycopg2>=2.7.7,<2.8.0
django-filter>=2.1.0,<2.2
django-extensions>=2.1.7,<2.2
python-dotenv>=0.9.1,<1.0
django-cors-headers>=3.1.0,<3.2
requests>=2.22.0,<2.23
django-debug-toolbar>=2.0,<2.1
python-dateutil>=2.8.1,<2.9
numpy>1.1,<2.0
drf-yasg>=1.17,<2.0
django-rest-passwordreset>=1.1.0,<1.2